backend/requirements.txt
langgraph
langchain
langchain-groq
langchain-community
PyGithub
python-dotenv
Flask
pymongo
requests
cryptography
PyJWT
gunicorn
pymongo[srv]